chore: change Message.id from i32 to i64
TDLib Message.id is i64, while the previous grammers-based implementation
used i32. Updating the domain type to i64 prevents potential data loss
during the TDLib migration.

Changes:
- domain::Message.id type changed from i32 to i64
- Updated grammers backend to convert i32 to i64
- Updated test helper functions in message_rendering, open_chat_state, shell
- Added #[allow(dead_code)] for unused items that will be used in Phase 5/6

+++ src/domain/message.rs
@@ -1,5 +1,6 @@
 /// Type of media attached to a message.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+#[allow(dead_code)] // Variants will be used in TDLib message mapper (Phase 5.3)
 pub enum MessageMedia {
     #[default]
     None,
@@ -40,7 +41,7 @@ impl MessageMedia {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Message {
-    pub id: i32,
+    pub id: i64,
     pub sender_name: String,
     pub text: String,
     pub timestamp_ms: i64,
