+++ src/usecases/shell.rs
@@ -350,7 +350,7 @@ mod tests {
         }
     }
 
-    fn message(id: i32, text: &str) -> Message {
+    fn message(id: i64, text: &str) -> Message {
         Message {
             id,
             sender_name: "User".to_owned(),
